services/sms_notify/test/subscribe/requirements.txt
aiohttp
asyncio-nats-client
